moved the helper math methods into utils
